feat: finally working crud tests

+++ OrderService/ProductService/Products/Product.cs
@@ -5,10 +5,10 @@ namespace ProductService.Products
 {
     public class Product
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public string Name { get; set; }
         [JsonIgnore]
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
         public Category Category { get; set; }
     }
 }
